docker/test_virus/requirements.txt
python-telegram-bot
telegram
pyyaml
python-dotenv
asyncio
